requirements.txt
torch
torchvision
triton

transformers==4.57.6
accelerate
datasets
safetensors
sentencepiece

peft==0.17.1
trl
bitsandbytes
flash-attn

qwen-vl-utils
byaldi
colpali-engine
mteb
sentence-transformers

pandas
numpy
scikit-learn
scipy

matplotlib
pillow
tqdm
rich

jsonlines
pyarrow
pdf2image
pytrec-eval-terrier

huggingface-hub==0.36.2
trackio

requests
python-dateutil
pytz
tzdata
